+++ Extend/Model/Payment.php
@@ -14,7 +14,7 @@ use OxidEsales\Eshop\Core\Exception\SystemComponentException;
 use Wirecard\Oxid\Core\OxidEeEvents;
 use Wirecard\Oxid\Core\PaymentMethodFactory;
 use Wirecard\Oxid\Model\MetaDataModel;
-use Wirecard\Oxid\Model\PaymentMethod;
+use Wirecard\Oxid\Model\PaymentMethod\PaymentMethod;
 
 /**
  * Extends the Payment model.
