+++ Model/PaymentMethod/BasePoiPiaPaymentMethod.php
@@ -7,71 +7,43 @@
  * https://github.com/wirecard/oxid-ee/blob/master/LICENSE
  */
 
-namespace Wirecard\Oxid\Model;
-
-use OxidEsales\Eshop\Core\Registry;
+namespace Wirecard\Oxid\Model\PaymentMethod;
 
 use Wirecard\Oxid\Core\Helper;
 use Wirecard\PaymentSdk\Config\Config;
 use Wirecard\PaymentSdk\Config\PaymentMethodConfig;
-use Wirecard\PaymentSdk\Entity\IdealBic;
-use Wirecard\PaymentSdk\Transaction\IdealTransaction;
+use Wirecard\PaymentSdk\Transaction\PoiPiaTransaction;
 use Wirecard\PaymentSdk\Transaction\Transaction;
 
 /**
- * Payment method implementation for iDEAL
+ * Base class for implementation of Payment on Invoice/Payment in Advance
  *
- * @since 1.2.0
+ * @since 1.3.0
  */
-class IdealPaymentMethod extends SepaCreditTransferPaymentMethod
+abstract class BasePoiPiaPaymentMethod extends SepaCreditTransferPaymentMethod
 {
-    /**
-     * @inheritdoc
-     *
-     * @since 1.2.0
-     */
-    protected static $_sName = "ideal";
-
     /**
      * @inheritdoc
      *
      * @var bool
      *
-     * @since 1.2.0
+     * @since 1.3.0
      */
     protected static $_bMerchantOnly = false;
-
-    /**
-     *
-     * @since 1.2.0
-     */
-    private $_aBankOptions = [];
-
-    /**
-     * IdealPaymentMethod constructor.
-     *
-     * @since 1.2.0
-     */
-    public function __construct()
-    {
-        parent::__construct();
-
-        $this->_aBankOptions = $this->getBanks();
-    }
 
     /**
      * @inheritdoc
      *
      * @return Config
      *
-     * @since 1.2.0
+     * @since 1.3.0
      */
     public function getConfig()
     {
         $oConfig = parent::getConfig();
 
         $oPaymentMethodConfig = new PaymentMethodConfig(
-            IdealTransaction::NAME,
+            PoiPiaTransaction::NAME,
             $this->_oPayment->oxpayments__wdoxidee_maid->value,
             $this->_oPayment->oxpayments__wdoxidee_secret->value
         );
@@ -86,11 +58,11 @@ class IdealPaymentMethod extends SepaCreditTransferPaymentMethod
      *
      * @return Transaction
      *
-     * @since 1.2.0
+     * @since 1.3.0
      */
     public function getTransaction()
     {
-        return new IdealTransaction();
+        return new PoiPiaTransaction();
     }
 
     /**
@@ -98,7 +70,7 @@ class IdealPaymentMethod extends SepaCreditTransferPaymentMethod
      *
      * @return array
      *
-     * @since 1.2.0
+     * @since 1.3.0
      */
     public function getConfigFields()
     {
@@ -153,39 +125,7 @@ class IdealPaymentMethod extends SepaCreditTransferPaymentMethod
      *
      * @return array
      *
-     * @since 1.2.0
-     */
-    public function getCheckoutFields()
-    {
-        return [
-            'bank' => [
-                'type' => 'select',
-                'options' => $this->_aBankOptions,
-                'title' => Helper::translate('wd_ideal_legend'),
-            ],
-        ];
-    }
-
-    /**
-     * Returns array for bank select options
-     *
-     * @return array
-     *
-     * @since 1.2.0
-     */
-    public function getBanks()
-    {
-        //IdealBic extends Enum class and contains const variables which represent bank options.
-        //toArray() combines all const variables from IdealBic and converts them to array.
-        return IdealBic::toArray();
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @return array
-     *
-     * @since 1.2.0
+     * @since 1.3.0
      */
     public function getPublicFieldNames()
     {
@@ -195,21 +135,5 @@ class IdealPaymentMethod extends SepaCreditTransferPaymentMethod
             'deleteCanceledOrder',
             'deleteFailedOrder',
         ]);
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @param IdealTransaction $oTransaction
-     * @param Order            $oOrder
-     *
-     * @since 1.2.0
-     */
-    public function addMandatoryTransactionData(&$oTransaction, $oOrder)
-    {
-        $oSession = Registry::getConfig()->getSession();
-        $aDynvalues = $oSession->getVariable('dynvalue');
-
-        $oTransaction->setBic($this->_aBankOptions[$aDynvalues['bank']]);
     }
 }
