+++ Model/PaymentMethod/InvoicePaymentMethod.php
@@ -7,7 +7,7 @@
  * https://github.com/wirecard/oxid-ee/blob/master/LICENSE
  */
 
-namespace Wirecard\Oxid\Model;
+namespace Wirecard\Oxid\Model\PaymentMethod;
 
 use OxidEsales\Eshop\Application\Model\Country;
 use OxidEsales\Eshop\Core\Exception\InputException;
